refactor(i18n): 将翻译逻辑提取为独立组件以提升可维护性
